PlataformaTURBIDEZ/backend/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.20
asyncpg>=0.28.0
geoalchemy2>=0.14.1
pydantic>=2.3.0
python-multipart>=0.0.6
geopandas>=0.14.0
numpy>=1.26.0
shapely>=2.0.1
openpyxl>=3.1.2
